# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Remote Patient Monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login for providers and patients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients (optionally filtered by provider)
      security:
        - bearerAuth: []
      parameters:
        - name: providerId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: riskLevel
          in: query
          schema:
            type: string
            enum: [normal, warning, critical]
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientSummary"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientRequest"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{patientId}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDetail"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientRequest"
      responses:
        "200":
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
...
  /patients/{patientId}/alerts:
    get:
      operationId: getPatientAlerts
      tags: [alerts]
      summary: Get alerts for a patient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all alerts across all patients
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [warning, critical]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertWithPatient"
  /alerts/{alertId}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: Acknowledge an alert
      security:
        - bearerAuth: []
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{alertId}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve an alert
      security:
        - bearerAuth: []
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - token
        - user
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [provider, admin, patient]
      required:
        - id
        - email
        - name
        - role
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        conditions:
          type: array
          items:
            type: string
        providerId:
          type: integer
        deviceType:
          type: string
          enum: [apple_health, google_fit, fitbit, garmin, manual]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - createdAt
    PatientSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        age:
          type: integer
        conditions:
          type: array
          items:
            type: string
        riskLevel:
          type: string
          enum: [normal, warning, critical]
        latestVitals:
          $ref: "#/components/schemas/VitalsSnapshot"
        activeAlertCount:
          type: integer
        lastSeen:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - riskLevel
        - activeAlertCount
    PatientDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        dateOfBirth:
          type: string
          format: date
        age:
          type: integer
        gender:
          type: string
        conditions:
          type: array
          items:
            type: string
        providerId:
          type: integer
        deviceType:
          type: string
        riskLevel:
          type: string
          enum: [normal, warning, critical]
        latestVitals:
          $ref: "#/components/schemas/VitalsSnapshot"
        thresholds:
          $ref: "#/components/schemas/VitalThresholds"
        activeAlertCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - riskLevel
        - activeAlertCount
        - createdAt
    CreatePatientRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        conditions:
          type: array
          items:
            type: string
        providerId:
          type: integer
        deviceType:
          type: string
          enum: [apple_health, google_fit, fitbit, garmin, manual]
      required:
        - name
        - email
        - password
    UpdatePatientRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        conditions:
          type: array
          items:
            type: string
        providerId:
          type: integer
        deviceType:
          type: string
          enum: [apple_health, google_fit, fitbit, garmin, manual]
...
    VitalsSnapshot:
      type: object
      properties:
        heartRate:
          type: number
        systolicBp:
          type: number
        diastolicBp:
          type: number
        spo2:
          type: number
        caloriesBurned:
          type: number
        temperature:
          type: number
        recordedAt:
          type: string
          format: date-time
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        vitalType:
          type: string
          enum: [heart_rate, blood_pressure, spo2, temperature]
        severity:
          type: string
          enum: [warning, critical]
        message:
          type: string
        value:
          type: number
        threshold:
          type: number
        status:
          type: string
          enum: [active, acknowledged, resolved]
        suggestedAction:
          type: string
        triggeredAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - vitalType
        - severity
        - message
        - status
        - triggeredAt
    AlertWithPatient:
      allOf:
        - $ref: "#/components/schemas/Alert"
        - type: object
          properties:
            patientName:
              type: string
            patientEmail:
              type: string
          required:
            - patientName
            - patientEmail
    VitalThresholds:
      type: object
      properties:
        patientId:
          type: integer
        heartRateMin:
          type: number
          default: 50
        heartRateMax:
          type: number
          default: 100
        heartRateCriticalMin:
          type: number
          default: 40
        heartRateCriticalMax:
          type: number
          default: 120
        systolicBpMin:
          type: number
          default: 90
        systolicBpMax:
          type: number
          default: 140
        systolicBpCriticalMin:
          type: number
          default: 80
        systolicBpCriticalMax:
          type: number
          default: 180
        diastolicBpMin:
          type: number
          default: 60
        diastolicBpMax:
          type: number
          default: 90
        diastolicBpCriticalMax:
          type: number
          default: 120
        spo2Min:
          type: number
          default: 95
        spo2CriticalMin:
          type: number
          default: 90
        temperatureMin:
          type: number
          default: 36.0
        temperatureMax:
          type: number
          default: 37.5
        temperatureCriticalMax:
          type: number
          default: 39.0
      required:
        - patientId
...