# lib/api-zod/src/generated/types/alertWithPatient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Remote Patient Monitoring API
 * OpenAPI spec version: 1.0.0
 */
import type { Alert } from "./alert";

export type AlertWithPatient = Alert & {
  patientName: string;
  patientEmail: string;
};