# lib/api-zod/src/generated/types/vitalThresholds.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Remote Patient Monitoring API
 * OpenAPI spec version: 1.0.0
 */

export interface VitalThresholds {
  patientId: number;
  heartRateMin?: number;
  heartRateMax?: number;
  heartRateCriticalMin?: number;
  heartRateCriticalMax?: number;
  systolicBpMin?: number;
  systolicBpMax?: number;
  systolicBpCriticalMin?: number;
  systolicBpCriticalMax?: number;
  diastolicBpMin?: number;
  diastolicBpMax?: number;
  diastolicBpCriticalMax?: number;
  spo2Min?: number;
  spo2CriticalMin?: number;
  temperatureMin?: number;
  temperatureMax?: number;
  temperatureCriticalMax?: number;
}