# lib/api-zod/src/generated/types/vitalsSnapshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Remote Patient Monitoring API
 * OpenAPI spec version: 1.0.0
 */

export interface VitalsSnapshot {
  heartRate?: number;
  systolicBp?: number;
  diastolicBp?: number;
  spo2?: number;
  caloriesBurned?: number;
  temperature?: number;
  recordedAt?: Date;
}